+++ include/renderer.h
@@ -16,8 +16,12 @@ void Render(EntityID ent, const glm::mat4& view, const glm::mat4& projection);
 // Renders an entity's b2Body
 void RenderBody(EntityID ent, const glm::mat4& view, const glm::mat4& projection);
 
+// Rendering objects for OpenGL
 inline unsigned int VAO, VBO, EBO;
+
+// Default 2D shader
 inline Shader defaultShader;
+// Shader for rendering green lines
 inline Shader lineShader;
 
 }