+++ src/camera.cpp
@@ -93,3 +93,35 @@ void Camera::updateCameraVectors()
     Right = glm::normalize(glm::cross(Front, WorldUp));
     Up = glm::normalize(glm::cross(Right, Front));
 }
+
+glm::vec3 Camera::ScreenToWorld(const glm::ivec2& pos)
+{
+    glm::vec2 viewportSize(Utils::globalWindow->width, Utils::globalWindow->height);
+
+    float zDepth = Position.z * 1.87f;
+
+    glm::mat4 projection = glm::perspective(glm::radians(GetZoom()), Utils::globalWindow->GetAspectRatio(), 0.1f, 100.0f);
+    glm::mat projInverse = glm::inverse(projection);
+
+    float mouse_x = pos.x;
+    float mouse_y = pos.y;
+
+    float ndc_x = (2.0f * mouse_x) / Utils::globalWindow->width - 1.0f;
+    float ndc_y = 1.0f - (2.0f * mouse_y) / Utils::globalWindow->height;
+
+    double focal_length = 1.0f / glm::tan(glm::radians(45.0f / 2.0f));
+    float ar = (float)Utils::globalWindow->height / (float)Utils::globalWindow->width;
+    glm::vec3 ray_view(ndc_x / focal_length, (ndc_y * ar) / focal_length, 1.0f);
+
+    glm::vec4 ray_ndc_4d(ndc_x, ndc_y, 1.0f, 1.0f);
+    glm::vec4 ray_view_4d = projInverse * ray_ndc_4d;
+
+    glm::vec4 view_space_intersect = glm::vec4(ray_view * zDepth, 1.0f);
+
+    glm::mat4 view = GetViewMatrix();
+    glm::mat4 viewInverse = glm::inverse(view);
+
+    glm::vec4 point_world = viewInverse * view_space_intersect;
+    
+    return glm::vec3(point_world);
+}